+++ test/PrimitiveInventorySpec.hs
@@ -46,6 +46,7 @@ spec = describe "MLF.Primitive.Inventory" $ do
           PrimitiveInventory.stringAppendPrimitiveName,
           PrimitiveInventory.stringReplaceCharPrimitiveName,
           PrimitiveInventory.stringIndexOfCharPrimitiveName,
+          PrimitiveInventory.stringIndexOfPrimitiveName,
           PrimitiveInventory.stringFromCharPrimitiveName,
           PrimitiveInventory.stringFromIntPrimitiveName,
           PrimitiveInventory.stringFromBoolPrimitiveName,
@@ -94,6 +95,8 @@ spec = describe "MLF.Primitive.Inventory" $ do
           name `shouldBe` PrimitiveInventory.stringReplaceCharPrimitiveName
         PrimitiveInventory.PrimitiveNativeStringIndexOfChar ->
           name `shouldBe` PrimitiveInventory.stringIndexOfCharPrimitiveName
+        PrimitiveInventory.PrimitiveNativeStringIndexOf ->
+          name `shouldBe` PrimitiveInventory.stringIndexOfPrimitiveName
         PrimitiveInventory.PrimitiveNativeStringFromChar ->
           name `shouldBe` PrimitiveInventory.stringFromCharPrimitiveName
         PrimitiveInventory.PrimitiveNativeStringFromInt ->